Resolved conflicts when renamed SelectFieldsPredicate to SelectSomeFieldsPredicate and SelectAllPredicate to SelectAllFieldsPredicate
