+++ textdb/textdb-textql/src/main/java/edu/uci/ics/textdb/textql/statements/predicates/SelectAllFieldsPredicate.java
@@ -1,5 +1,8 @@
 package edu.uci.ics.textdb.textql.statements.predicates;
 
+import edu.uci.ics.textdb.textql.planbuilder.beans.PassThroughBean;
+import edu.uci.ics.textdb.web.request.beans.OperatorBean;
+
 /**
  * Object representation of a "SELECT *" predicate inside a { @code SelectExtractStatement }.
  * 
@@ -8,6 +11,15 @@ package edu.uci.ics.textdb.textql.statements.predicates;
  */
 public class SelectAllFieldsPredicate extends SelectPredicate {
 
+    /**
+     * Return this operator converted to an { @code OperatorBean }.
+     * @param selectOperatorId The ID of the OperatorBean to be created.
+     */
+    public OperatorBean getOperatorBean(String selectOperatorId) {
+        return new PassThroughBean(selectOperatorId, "PassThrough");
+    }
+    
+    
     @Override
     public boolean equals(Object other) {
         if (other == null) { return false; }
